git plugin: modified extraction scheme to avoid detached head.
In the previous ectraction scheme, the branch was ignored, creating
a detached head when a revision is specified.
The nez scheme ill first extract the branch and then reset it to the
given revision, avoiding this effect.
This change also maked optional a number of git fields, actually:
alias, label, revision are now optional.