deptool.py: removed unsafe coupling with plugin in dump...() functions.
The dump...() function were badly coupled with the plugins with a
sharing of the repository map. If the map was changed on one side
it may have unexpected effect.
The plugins now provide an access to the name of the component with
the name() method.
The generation of the manifest in the dump...() function is now
done by a deep copy of the initial manifest and a modification
of the revision before the dump.