git plugin: implement local git cache for performance
Change git plugin repository management to transparently
use a local git repository cache.

This allow a user to remove an element of the manifest
with rm -rf for instance and run again extract without the
cost of fetching the whole repository again.

The repository cache is a bare repository that fetches heads and
tags from the remote. It is created in a local cache dir for the
plugin (.deptools/cache/plugins/git/) with the path
<guid_path>/<base>.git. The <guid_path> is a globally unique
path identifier for the repository URI and <base> is the
basename of the repository URI.
The <guid_path> has the format of git objects path, i.e.
sha1[:2] + '/' + sha1[2:] where sha1 is the sha1 sum of the
repository URI string.

On extraction the actual repository is initialized with
--reference <path_to_cached_repo> such that the repository does
not fetch objects already present in the cached repository.

All operations that need to access the remote must first
fetch the cached repository (extract, update, rebase, deliver).

Change-Id: I6c591e67bcdf553be0c1300b1cdaee8b1f200685